requirements.txt
django
gunicorn
psycopg2
celery
pytz
djangorestframework
djangorestframework-simplejwt
django-ratelimit
redis
hiredis